DT-1769 Split cypress spec into 3 small test suites
